Stemming and lemmatization are both text normalization techniques used in (NLP) to reduce words to their base or root form.
Stemming and lemmatization are both text normalization techniques used in (NLP) to reduce words to their base or root form. The purpose of these processes is to treat different forms of a word as the same word, which can be particularly useful in tasks like text mining, search engines, and information retrieval.